v2/requirements.txt
# VLM Server - Python FastAPI Backend
# Python 3.9+

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Image Processing
opencv-python==4.12.0.88
pillow==11.3.0

# AI/ML - Core
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0
transformers==4.57.1
accelerate==1.10.1

# AI/ML - Vision Language Models
safetensors==0.6.2
peft==0.17.1
trl==0.24.0
timm==1.0.22

# Object Detection (YOLO)
ultralytics==8.3.67

# Data Processing
numpy==2.0.2
pandas==2.3.3
datasets==4.4.1

# API & HTTP
requests==2.32.5
requests-toolbelt==1.0.0
httpx==0.28.1
httpx-sse==0.4.3
aiohttp==3.13.2

# Configuration & Environment
python-dotenv==1.2.1
pydantic==2.12.4
pydantic-settings==2.11.0

# Utilities
typing-extensions==4.15.0
click==8.1.8
tqdm==4.67.1
psutil==7.1.3

# CORS Support
flask-cors==6.0.1

